# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: AddrPass API
  description: Privacy-first address management API. Store addresses, share via tokenized links and QR codes.
  version: 0.1.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  contact:
    url: https://github.com/addrpass/addrpass
servers:
  - url: https://api.addrpass.com
    description: Production
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/shares/{id}/accesses:
    get:
      summary: Get access logs
      operationId: getAccessLogs
      tags: [Shares]
      security: [{ bearer: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Access log list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccessLog"
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AccessLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        share_id:
          type: string
          format: uuid
        ip:
          type: string
        user_agent:
          type: string
        country:
          type: string
        access_at:
          type: string
          format: date-time